More components converted to TS
